Edit the tests, free the data in the onplace method, other stuffs

+++ Sub.hpp
@@ -9,7 +9,10 @@
 class Sub : public Operator {
 public:
     size_t apply(size_t leftValue, size_t rightValue, size_t modulo) const {
-        long int result = leftValue - rightValue;
+        // Stock a resilt in a long long to fit for a size_t
+        long long result = leftValue - rightValue;
+
+        // Get a real modulo. Example : -1 % 5 = 0, but it should be 4
         return (((result % (int) modulo) + modulo) % modulo);
     }
 };
